requirements.txt
certifi==2018.4.16
future==0.16.0
PyMySQL==0.8.1
python-decouple==3.1
python-telegram-bot==10.1.0
requests==2.19.1